src/services/user/aws_lambdas/requirements.txt
boto3>=1.26.0
pyjwt>=2.8.0